+++ tests/testthat/test-checkArgumentsPDTMTptm.R
@@ -27,9 +27,6 @@ test_that("argument checking for PD-TMT PTM works", {
                                            splitSeparator = ";", joinSeparator = "."),
         proteinIdColProteins = "einprotProtein",
         proteinIdColPeptides = "einprotProtein",
-        modificationsCol = "Modifications.in.Master.Proteins",
-        excludeUnmodifiedPeptides = FALSE,
-        keepModifications = NULL,
         comparisons = list(),
         ctrlGroup = "",
         allPairwiseComparisons = TRUE,
@@ -198,30 +195,6 @@ test_that("argument checking for PD-TMT PTM works", {
     args$proteinIdColPeptides <- function(x, y) x + 1
     expect_error(do.call(.checkArgumentsPDTMTptm, args),
                  "length(formals(proteinIdColPeptides)) == 1 is not TRUE", fixed = TRUE)
-
-    ## modificationsCol
-    args <- args0
-    args$modificationsCol <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'modificationsCol' must be of class 'character'")
-    args$modificationsCol <- c("Modifications", "Modifications.in.Master.Proteins")
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'modificationsCol' must have length 1")
-
-    ## excludeUnmodifiedPeptides
-    args <- args0
-    args$excludeUnmodifiedPeptides <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'excludeUnmodifiedPeptides' must be of class 'logical'")
-    args$excludeUnmodifiedPeptides <- c(TRUE, FALSE)
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'excludeUnmodifiedPeptides' must have length 1")
-
-    ## keepModifications
-    args <- args0
-    args$keepModifications <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'keepModifications' must be of class 'character'")
 
     ## comparisons
     args <- args0