+++ tests/testthat/test-runMaxQuantAnalysis.R
@@ -25,6 +25,8 @@ test_that("runMaxQuantAnalysis works", {
                                             separator = ";"),
         proteinIdCol = function(df) getFirstId(df, colName = "Majority.protein.IDs",
                                                separator = ";"),
+        stringIdCol = function(df) combineIds(df, combineCols = c("Gene.names", "Majority.protein.IDs"),
+                                              combineWhen = "missing", makeUnique = FALSE),
         iColPattern = "^iBAQ\\\\.",
         sampleAnnot = data.frame(
             sample = c("Adnp_IP04", "Adnp_IP05", "Adnp_IP06",
@@ -209,6 +211,12 @@ test_that("runMaxQuantAnalysis works", {
     args$proteinIdCol <- 1
     expect_error(do.call(runMaxQuantAnalysis, args),
                  "'proteinIdCol' must be of class 'character'")
+
+    ## stringIdCol
+    args <- args0
+    args$stringIdCol <- 1
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'stringIdCol' must be of class 'character'")
 
     ## iColPattern
     args <- args0